Check if buffer opened by cscope was already open
Previous fix for early buffer closing relied on saving and later
comparing current file name. This resulted in unwanted buffer closing
when the file, that cscope jumped to, was already opened before.

Current version saves IDs of all currently opened buffers and later on
checks if the ID of buffer opened by cscope was in previously opened
buffer list.